refactor(issue-95): split repo policy parsing and tests
