+++ python_tools/policies/repo_policy.py
@@ -1,51 +1,28 @@
 #!/usr/bin/env python3
 from __future__ import annotations
 
-import re
 from pathlib import Path
 
 from python_tools.core.base_check import BaseCheck
 from python_tools.core.models import CheckContext, CheckResult
 from python_tools.policies.header_definition_policy import find_header_function_definition_lines
-from python_tools.policies.repo_policy_power_of_10 import check_power_of_10_content
+from python_tools.policies.repo_policy_support import (
+    CPP_SCAN_EXTS,
+    EVIDENCE_WORKFLOW_PATHS,
+    FORBIDDEN_CPP_EXTS,
+    FORBIDDEN_MARKER_RE,
+    HEADER_EXTS,
+    LEGACY_CTEST_SELECTORS,
+    LINE_COUNT_NAMES,
+    TEXT_EXTS,
+    check_cpp_content,
+    collect_marker_commands,
+    load_allowlist,
+    should_count_lines,
+    should_skip_dir,
+)
 
-FORBIDDEN_CPP_EXTS = {".h", ".hh", ".hxx", ".c", ".cc", ".cxx"}
-LINE_COUNT_EXTS = {".cmake", ".cpp", ".cu", ".cuh", ".h", ".hh", ".hpp", ".hxx", ".inl", ".ini", ".json", ".md", ".mk", ".ps1", ".py", ".sh", ".toml", ".txt", ".xml", ".yaml", ".yml"}
-LINE_COUNT_NAMES = {"CMakeLists.txt", "Makefile"}
-IGNORED_DIRS = {".git", ".idea", ".vs", "__pycache__", "build", "exports"}
-HEADER_EXTS = {".hpp", ".h", ".hh", ".hxx"}
-TEXT_EXTS = LINE_COUNT_EXTS | {".json", ".toml", ".xml"}
-CPP_SCAN_EXTS = {".cpp", ".cc", ".cxx", ".c", ".cu", ".cuh", ".hpp", ".h", ".hh", ".hxx", ".inl"}
-PROD_ROOTS = ("apps/", "engine/", "runtime/", "modules/")
-EVIDENCE_WORKFLOW_PATHS = (".github/workflows/pr-fast-quality-gate.yml", ".github/workflows/nightly-full.yml", ".github/workflows/release-lane.yml")
-LEGACY_CTEST_SELECTORS = ("ConfigArgsTest", "BackendProtocolTest", "FrontendBridgeTest", "FrontendRuntimeTest", "QtMainWindowTest")
-FORBIDDEN_MARKER_RE = re.compile(r"(?m)(#|//|/\*)\s*(TODO|FIXME|HACK)\b")
-GTEST_INCLUDE_RE = re.compile(r'#include\s*[<"]gtest/gtest\.h[>"]')
-GTEST_MACRO_RE = re.compile(r"\bTEST(?:_F)?\s*\(")
-UNNAMED_NAMESPACE_RE = re.compile(r"(?m)^\s*namespace\s*\{")
-USING_ANY_RE = re.compile(r"(?m)^\s*using\b[^;]*;")
-INLINE_NAMESPACE_RE = re.compile(r"(?m)^\s*namespace\s+[A-Za-z0-9_]+::[A-Za-z0-9_:]+\s*\{")
-NAMESPACE_BLOCK_RE = re.compile(r"(?m)^\s*namespace\s+[A-Za-z0-9_]+\s*\{")
-GRAVITY_INTERNAL_NAMESPACE_RE = re.compile(r"(?m)^\s*namespace\s+gravity_internal_")
-QT_REFERENCE_NEW_RE = re.compile(r"(?m)^\s*(?:auto|Q[A-Za-z0-9_<>:]+)\s*&\s*[A-Za-z0-9_]+\s*=\s*\*new\s+Q[A-Za-z0-9_<>:]+\s*\(")
-PREPROCESSOR_CONDITIONAL_RE = re.compile(r"(?m)^\s*#(?:if|ifdef|ifndef|elif|else|endif)\b")
-PRAGMA_ONCE_RE = re.compile(r"(?m)^\s*#pragma\s+once\b")
-DEFINE_RE = re.compile(r"(?m)^\s*#define\s+([A-Z][A-Z0-9_]+)\b(?!\s*\()")
-def should_skip_dir(dirname: str) -> bool:
-    return dirname in IGNORED_DIRS or dirname.startswith(("build-", "cmake-build-", ".pytest-basetemp"))
-def should_count_lines(path: Path) -> bool:
-    return path.name in LINE_COUNT_NAMES or path.suffix.lower() in LINE_COUNT_EXTS
-def load_allowlist(path: Path) -> set[str]:
-    if not path.exists():
-        return set()
-    entries: set[str] = set()
-    for raw in path.read_text(encoding="utf-8").splitlines():
-        line = raw.strip()
-        if line and not line.startswith("#"):
-            entries.add(line)
-    return entries
-def is_prod_path(rel: str) -> bool:
-    return rel.startswith(PROD_ROOTS)
+
 class RepoPolicyCheck(BaseCheck):
     name = "repo"
     success_message = "Repository policy check passed"
@@ -68,8 +45,6 @@ class RepoPolicyCheck(BaseCheck):
                 result.add_error(f"{rel}: forbidden C/C++ extension '{suffix}' (use .hpp/.cpp)")
             if suffix in TEXT_EXTS or path.name in LINE_COUNT_NAMES:
                 content = path.read_text(encoding="utf-8", errors="ignore")
-                if suffix in HEADER_EXTS and USING_ANY_RE.search(content):
-                    result.add_error(f"{rel}: forbidden 'using' in header")
                 if suffix in HEADER_EXTS:
                     for line_number in find_header_function_definition_lines(content):
                         result.add_error(f"{rel}:{line_number}: function definitions in headers are forbidden")
@@ -77,7 +52,7 @@ class RepoPolicyCheck(BaseCheck):
                 if marker_match:
                     result.add_error(f"{rel}: forbidden marker '{marker_match.group(2)}' found")
                 if suffix in CPP_SCAN_EXTS:
-                    self._check_cpp_content(rel, content, result)
+                    check_cpp_content(rel, content, result)
             if not should_count_lines(path):
                 continue
             if content is None:
@@ -88,60 +63,6 @@ class RepoPolicyCheck(BaseCheck):
         self._check_evidence_workflow_commands(context.root, result, "cmake -S", "-DGRAVITY_PROFILE=prod", "evidence configure command must include -DGRAVITY_PROFILE=prod")
         self._check_evidence_workflow_commands(context.root, result, "ctest ", "--no-tests=error", "CI ctest command must include --no-tests=error")
         self._check_legacy_ctest_selectors(context.root, result)
-
-    def _check_cpp_content(self, rel: str, content: str, result: CheckResult) -> None:
-        suffix = Path(rel).suffix.lower()
-        if not rel.startswith("tests/"):
-            if GTEST_INCLUDE_RE.search(content):
-                result.add_error(f"{rel}: gtest include found outside tests/")
-            if GTEST_MACRO_RE.search(content):
-                result.add_error(f"{rel}: gtest TEST macro found outside tests/")
-        if UNNAMED_NAMESPACE_RE.search(content):
-            detail = "unnamed namespace is forbidden in production paths" if is_prod_path(rel) else "unnamed namespace is forbidden"
-            result.add_error(f"{rel}: {detail}")
-        if USING_ANY_RE.search(content):
-            result.add_error(f"{rel}: 'using' is forbidden in C++ sources")
-        if INLINE_NAMESPACE_RE.search(content):
-            result.add_error(f"{rel}: nested namespace declaration (A::B) is forbidden")
-        if GRAVITY_INTERNAL_NAMESPACE_RE.search(content):
-            result.add_error(f"{rel}: gravity_internal_* namespace is forbidden")
-        if is_prod_path(rel) and len(NAMESPACE_BLOCK_RE.findall(content)) > 1:
-            result.add_error(f"{rel}: nested namespace blocks are forbidden in production paths")
-        if PRAGMA_ONCE_RE.search(content):
-            result.add_error(f"{rel}: #pragma once is forbidden; use include guards")
-        if suffix in HEADER_EXTS:
-            self._check_include_guard(rel, content, result)
-        else:
-            if PREPROCESSOR_CONDITIONAL_RE.search(content):
-                result.add_error(f"{rel}: preprocessor conditionals are forbidden in C/C++ sources")
-            if DEFINE_RE.search(content):
-                result.add_error(f"{rel}: preprocessor macros are forbidden in C/C++ sources")
-        if is_prod_path(rel):
-            for error in check_power_of_10_content(rel, content):
-                result.add_error(error)
-        if rel.startswith("modules/qt/") and QT_REFERENCE_NEW_RE.search(content):
-            result.add_error(f"{rel}: Qt '*new + reference' ownership pattern is forbidden")
-
-    def _check_include_guard(self, rel: str, content: str, result: CheckResult) -> None:
-        non_empty = [(index, line.strip()) for index, line in enumerate(content.splitlines()) if line.strip()]
-        if len(non_empty) < 3:
-            result.add_error(f"{rel}: header must use a strict include guard")
-            return
-        (_, first), (_, second), (_, last) = non_empty[0], non_empty[1], non_empty[-1]
-        if not first.startswith("#ifndef "):
-            result.add_error(f"{rel}: header must start with #ifndef include guard")
-            return
-        guard_name = first.split(maxsplit=1)[1]
-        if second != f"#define {guard_name}":
-            result.add_error(f"{rel}: header include guard must define the same symbol on the second line")
-        if not last.startswith("#endif"):
-            result.add_error(f"{rel}: header must end with #endif include guard")
-        conditional_positions = [index for index, line in non_empty if PREPROCESSOR_CONDITIONAL_RE.match(line)]
-        if conditional_positions != [non_empty[0][0], non_empty[-1][0]]:
-            result.add_error(f"{rel}: header must not use preprocessor conditionals beyond the include guard")
-        define_positions = [(index, name) for index, line in non_empty for name in DEFINE_RE.findall(line)]
-        if define_positions != [(non_empty[1][0], guard_name)]:
-            result.add_error(f"{rel}: header must not define macros beyond the include guard")
 
     def _check_line_count(self, rel: str, content: str, context: CheckContext, allowlist: set[str], used_allowlist: set[str], result: CheckResult) -> None:
         line_count = len(content.splitlines())
@@ -160,27 +81,9 @@ class RepoPolicyCheck(BaseCheck):
             if not path.exists():
                 continue
             content = path.read_text(encoding="utf-8", errors="ignore")
-            for command in self._collect_marker_commands(content, marker_text):
+            for command in collect_marker_commands(content, marker_text):
                 if required_text not in command:
                     result.add_error(f"{rel}: {error_text}: {command}")
-
-    def _collect_marker_commands(self, content: str, marker_text: str) -> list[str]:
-        commands: list[str] = []
-        lines = content.splitlines()
-        index = 0
-        while index < len(lines):
-            stripped = lines[index].strip()
-            marker = stripped.find(marker_text)
-            if marker == -1:
-                index += 1
-                continue
-            command = stripped[marker:]
-            while command.endswith("\\") and index + 1 < len(lines):
-                index += 1
-                command = f"{command} {lines[index].strip()}"
-            commands.append(command)
-            index += 1
-        return commands
 
     def _check_legacy_ctest_selectors(self, root: Path, result: CheckResult) -> None:
         for rel in EVIDENCE_WORKFLOW_PATHS:
@@ -188,6 +91,6 @@ class RepoPolicyCheck(BaseCheck):
             if not path.exists():
                 continue
             content = path.read_text(encoding="utf-8", errors="ignore")
-            for command in self._collect_marker_commands(content, "ctest "):
+            for command in collect_marker_commands(content, "ctest "):
                 if any(selector in command for selector in LEGACY_CTEST_SELECTORS):
                     result.add_error(f"{rel}: CI ctest selector must use normalized TST_* ids: {command}")
