+++ tests/checks/suites/policy/test_repo_policy_headers.py
@@ -3,11 +3,11 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from tests.checks.suites.policy.test_repo_policy import _run, _write
+from tests.checks.suites.policy.repo_policy_test_support import run_repo_policy, write_file
 
 
 def test_repo_policy_rejects_function_definition_in_header(tmp_path: Path) -> None:
-    _write(
+    write_file(
         tmp_path / "runtime" / "include" / "protocol" / "bad.hpp",
         "class Bad {\n"
         "    public:\n"
@@ -16,33 +16,39 @@ def test_repo_policy_rejects_function_definition_in_header(tmp_path: Path) -> No
         "        }\n"
         "};\n",
     )
-    ok, errors, _ = _run(tmp_path, tmp_path / "allowlist.txt")
+    ok, errors, _ = run_repo_policy(tmp_path)
     assert not ok
     assert any("function definitions in headers are forbidden" in error for error in errors)
 
 
 def test_repo_policy_accepts_declaration_only_header(tmp_path: Path) -> None:
-    _write(
+    write_file(
         tmp_path / "runtime" / "include" / "protocol" / "good.hpp",
+        "#ifndef GRAVITY_RUNTIME_INCLUDE_PROTOCOL_GOOD_HPP_\n"
+        "#define GRAVITY_RUNTIME_INCLUDE_PROTOCOL_GOOD_HPP_\n"
         "class Good {\n"
         "    public:\n"
         "        int value() const;\n"
-        "};\n",
+        "};\n"
+        "#endif // GRAVITY_RUNTIME_INCLUDE_PROTOCOL_GOOD_HPP_\n",
     )
-    ok, errors, _ = _run(tmp_path, tmp_path / "allowlist.txt")
+    ok, errors, _ = run_repo_policy(tmp_path)
     assert ok
     assert not errors
 
 
 def test_repo_policy_accepts_header_declaration_with_default_braced_arg(tmp_path: Path) -> None:
-    _write(
+    write_file(
         tmp_path / "runtime" / "include" / "protocol" / "good_default.hpp",
+        "#ifndef GRAVITY_RUNTIME_INCLUDE_PROTOCOL_GOOD_DEFAULT_HPP_\n"
+        "#define GRAVITY_RUNTIME_INCLUDE_PROTOCOL_GOOD_DEFAULT_HPP_\n"
         "#include <functional>\n"
         "class GoodDefault {\n"
         "    public:\n"
         "        bool run(const std::function<void()> &callback = {});\n"
-        "};\n",
+        "};\n"
+        "#endif // GRAVITY_RUNTIME_INCLUDE_PROTOCOL_GOOD_DEFAULT_HPP_\n",
     )
-    ok, errors, _ = _run(tmp_path, tmp_path / "allowlist.txt")
+    ok, errors, _ = run_repo_policy(tmp_path)
     assert ok
     assert not errors
